implement level in SimpleLogger, remove LevelLogger struct
