--- logger/level_logger.go
@@ -1,14 +1,0 @@
-package logger
-
-import (
-	"context"
-)
-
-type LevelLogger struct {
-	SimpleLogger
-	Level Level
-}
-
-func (l *LevelLogger) Enabled(_ context.Context, level Level) bool {
-	return level >= l.Level
-}
